Create generic processing scene
This is a generic wrapper created from the swap processing scene to be reused for other long running operation and is cancellable after a set amount of time.
